requirements.txt
Mastodon.py==1.8.1
ics==0.7.2
requests==2.32.3
